% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{AggregateMovementsData}
\alias{AggregateMovementsData}
\title{Call the aggregatemovementsdata API}
\usage{
AggregateMovementsData(client, queryName)
}
\arguments{
\item{client}{The PetroLogistics API client}

\item{queryName}{The query_name parameter passed to the API}
}
\value{
list(success=T/F, error="error string"/NA, data=parsed_response/NA) 'success' indicates
  if the API request completed successfully. 'error' has the error string if the request was not
  successful. 'data' as the parsed response, if available. It has the general format
  list(envelope=list(header=list(..), movements=list(..)))
}
\description{
This function calls PetroLogistics' aggregatemovementsdata API and returns the parsed result
}
\examples{
AggregateMovementsData(client, "Angola_Test_Data")
}
