% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{PetroLogistics}
\alias{PetroLogistics}
\title{PetroLogistics API constructor}
\usage{
PetroLogistics(username, password, apiKey, apiHash)
}
\arguments{
\item{username}{Username}

\item{password}{Password}

\item{apiKey}{API Key}

\item{apiHash}{API Hash}
}
\description{
This function creates a client for the PetroLogistics API
}
